refactor: Flatten plugins/intraflow structure
Moves intra-flow policy implementations directly into the
`plugins/intraflow` package, removing the unnecessary nested
directories.

This simplifies the import paths and project structure.

This is a no-op refactoring.
