+++ pkg/epp/flowcontrol/framework/plugins/intraflow/fcfs_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package fcfs
+package intraflow
 
 import (
 	"testing"
